Refactored and optimized code of export, browser, pool and other modules.
