+++ lib/cache.js
@@ -390,11 +390,17 @@ export const checkAndUpdateCache = async (options) => {
 export const getCachePath = () =>
   join(__dirname, getOptions().highcharts.cachePath);
 
+export const getCache = () => cache;
+
+export const highcharts = () => cache.sources;
+
+export const version = () => cache.hcVersion;
+
 export default {
   checkAndUpdateCache,
   getCachePath,
   updateVersion,
-  getCache: () => cache,
-  highcharts: () => cache.sources,
-  version: () => cache.hcVersion
+  getCache,
+  highcharts,
+  version
 };
