+++ lib/export.js
@@ -12,18 +12,14 @@ See LICENSE file in root for details.
 
 *******************************************************************************/
 
-import { readFileSync } from 'fs';
-import path from 'path';
-import * as url from 'url';
-
-import cache from './cache.js';
+import { addPageResources, clearPageResources } from './browser.js';
+import { getCache } from './cache.js';
 import { triggerExport } from './highcharts.js';
-import { log, logWithStack } from './logger.js';
+import { log } from './logger.js';
+
 import svgTemplate from './../templates/svg_export/svg_export.js';
 
 import ExportError from './errors/ExportError.js';
-
-const __basedir = url.fileURLToPath(new URL('.', import.meta.url));
 
 /**
  * Retrieves the clipping region coordinates of the specified page element with
@@ -138,7 +134,7 @@ const createSVG = (page) =>
  *
  * @returns {Promise<void>} Promise resolving after the configuration is set.
  */
-const setAsConfig = (page, chart, options, displayErrors) =>
+const setAsConfig = async (page, chart, options, displayErrors) =>
   page.evaluate(triggerExport, chart, options, displayErrors);
 
 /**
@@ -152,55 +148,8 @@ const setAsConfig = (page, chart, options, displayErrors) =>
  * the exported data or rejecting with an ExportError.
  */
 export default async (page, chart, options) => {
-  /**
-   * Keeps track of all resources added on the page with addXXXTag. etc
-   * It's VITAL that all added resources ends up here so we can clear things
-   * out when doing a new export in the same page!
-   */
-  const injectedResources = [];
-
-  /** Clear out all state set on the page with addScriptTag/addStyleTag. */
-  const clearInjected = async (page) => {
-    for (const resource of injectedResources) {
-      await resource.dispose();
-    }
-
-    // Destroy old charts after export is done and reset all CSS and script tags
-    await page.evaluate(() => {
-      // We are not guaranteed that Highcharts is loaded, e,g, when doing SVG
-      // exports
-      if (typeof Highcharts !== 'undefined') {
-        // eslint-disable-next-line no-undef
-        const oldCharts = Highcharts.charts;
-
-        // Check in any already existing charts
-        if (Array.isArray(oldCharts) && oldCharts.length) {
-          // Destroy old charts
-          for (const oldChart of oldCharts) {
-            oldChart && oldChart.destroy();
-            // eslint-disable-next-line no-undef
-            Highcharts.charts.shift();
-          }
-        }
-      }
-
-      // eslint-disable-next-line no-undef
-      const [...scriptsToRemove] = document.getElementsByTagName('script');
-      // eslint-disable-next-line no-undef
-      const [, ...stylesToRemove] = document.getElementsByTagName('style');
-      // eslint-disable-next-line no-undef
-      const [...linksToRemove] = document.getElementsByTagName('link');
-
-      // Remove tags
-      for (const element of [
-        ...scriptsToRemove,
-        ...stylesToRemove,
-        ...linksToRemove
-      ]) {
-        element.remove();
-      }
-    });
-  };
+  // Injected resources array (additional JS and CSS)
+  let injectedResources = [];
 
   try {
     log(4, '[export] Determining export path.');
@@ -210,7 +159,7 @@ export default async (page, chart, options) => {
     // Decide whether display error or debbuger wrapper around it
     const displayErrors =
       exportOptions?.options?.chart?.displayErrors &&
-      cache.getCache().activeManifest.modules.debugger;
+      getCache().activeManifest.modules.debugger;
 
     let isSVG;
     if (
@@ -256,95 +205,10 @@ export default async (page, chart, options) => {
       }
     }
 
-    // Use resources
-    const resources = options.customLogic.resources;
-    if (resources) {
-      const injectedJs = [];
-
-      // Load custom JS code
-      if (resources.js) {
-        injectedJs.push({
-          content: resources.js
-        });
-      }
-
-      // Load scripts from all custom files
-      if (resources.files) {
-        for (const file of resources.files) {
-          const isLocal = !file.startsWith('http') ? true : false;
-
-          // Add each custom script from resources' files
-          injectedJs.push(
-            isLocal
-              ? {
-                  content: readFileSync(file, 'utf8')
-                }
-              : {
-                  url: file
-                }
-          );
-        }
-      }
-
-      for (const jsResource of injectedJs) {
-        try {
-          injectedResources.push(await page.addScriptTag(jsResource));
-        } catch (error) {
-          logWithStack(2, error, `[export] The JS resource cannot be loaded.`);
-        }
-      }
-      injectedJs.length = 0;
-
-      // Load CSS
-      const injectedCss = [];
-      if (resources.css) {
-        let cssImports = resources.css.match(/@import\s*([^;]*);/g);
-        if (cssImports) {
-          // Handle css section
-          for (let cssImportPath of cssImports) {
-            if (cssImportPath) {
-              cssImportPath = cssImportPath
-                .replace('url(', '')
-                .replace('@import', '')
-                .replace(/"/g, '')
-                .replace(/'/g, '')
-                .replace(/;/, '')
-                .replace(/\)/g, '')
-                .trim();
-
-              // Add each custom css from resources
-              if (cssImportPath.startsWith('http')) {
-                injectedCss.push({
-                  url: cssImportPath
-                });
-              } else if (options.customLogic.allowFileResources) {
-                injectedCss.push({
-                  path: path.join(__basedir, cssImportPath)
-                });
-              }
-            }
-          }
-        }
-
-        // The rest of the CSS section will be content by now
-        injectedCss.push({
-          content: resources.css.replace(/@import\s*([^;]*);/g, '') || ' '
-        });
-
-        for (const cssResource of injectedCss) {
-          try {
-            injectedResources.push(await page.addStyleTag(cssResource));
-          } catch (error) {
-            logWithStack(
-              2,
-              error,
-              `[export] The CSS resource cannot be loaded.`
-            );
-          }
-        }
-        injectedCss.length = 0;
-      }
-    }
+    // Keeps track of all resources added on the page with addXXXTag. etc
+    // It's VITAL that all added resources ends up here so we can clear things
+    // out when doing a new export in the same page!
+    injectedResources = await addPageResources(page, options);
 
     // Get the real chart size and set the zoom accordingly
     const size = isSVG
@@ -401,7 +265,7 @@ export default async (page, chart, options) => {
     });
 
     let data;
-    // RASTERIZATION
+    // Rasterization process
     if (exportOptions.type === 'svg') {
       // SVG
       data = await createSVG(page);
@@ -434,10 +298,11 @@ export default async (page, chart, options) => {
       );
     }
 
-    await clearInjected(page);
+    // Clear previously injected JS and CSS resources
+    await clearPageResources(page, injectedResources);
     return data;
   } catch (error) {
-    await clearInjected(page);
+    await clearPageResources(page, injectedResources);
     return error;
   }
 };
