+++ lib/highcharts.js
@@ -30,7 +30,7 @@ export function setupHighcharts() {
  * @param {object} options - The export options.
  * @param {boolean} displayErrors - A flag indicating whether to display errors.
  */
-export function triggerExport(chartOptions, options, displayErrors) {
+export async function triggerExport(chartOptions, options, displayErrors) {
   // Display errors flag taken from chart options nad debugger module
   window._displayErrors = displayErrors;
