+++ lib/pool.js
@@ -16,12 +16,11 @@ import { Pool } from 'tarn';
 import { v4 as uuid } from 'uuid';
 
 import {
-  close as browserClose,
   create as createBrowser,
-  newPage as browserNewPage,
+  close as closeBrowser,
+  newPage,
   clearPage
 } from './browser.js';
-import { getOptions } from './config.js';
 import puppeteerExport from './export.js';
 import { log, logWithStack } from './logger.js';
 import { measureTime } from './utils.js';
@@ -60,7 +59,7 @@ const factory = {
     const startDate = new Date().getTime();
 
     try {
-      page = await browserNewPage();
+      page = await newPage();
 
       if (!page || page.isClosed()) {
         throw new ExportError('The page is invalid or closed.');
@@ -77,30 +76,6 @@ const factory = {
         'Error encountered when creating a new page.'
       ).setError(error);
     }
-
-    const { debug } = getOptions();
-    // Set the console listener, if needed
-    if (debug.enable && debug.listenToConsole) {
-      page.on('console', (message) => {
-        console.log(`[debug] ${message.text()}`);
-      });
-    }
-
-    // Set the pageerror listener
-    page.on('pageerror', async (error) => {
-      // TODO: Consider adding a switch here that turns on log(0) logging
-      // on page errors.
-      await page.$eval(
-        '#container',
-        (element, errorMessage) => {
-          // eslint-disable-next-line no-undef
-          if (window._displayErrors) {
-            element.innerHTML = errorMessage;
-          }
-        },
-        `<h1>Chart input data error: </h1>${error.toString()}`
-      );
-    });
 
     return {
       id,
@@ -259,7 +234,7 @@ export async function killPool() {
   }
 
   // Close the browser instance
-  await browserClose();
+  await closeBrowser();
 }
 
 /**
@@ -336,7 +311,7 @@ export const postWork = async (chart, options) => {
       // TODO: If the export failed because puppeteer timed out, we need to force kill the worker so we get a new page. That needs to be handled better than this hack.
       if (result.message === 'Rasterization timeout') {
         workerHandle.page.close();
-        workerHandle.page = await browserNewPage();
+        workerHandle.page = await newPage();
       }
 
       throw new ExportError(
