+++ lib/server/routes/change_hc_version.js
@@ -12,7 +12,7 @@ See LICENSE file in root for details.
 
 *******************************************************************************/
 
-import cache from '../../cache.js';
+import { updateVersion, version } from '../../cache.js';
 import { envs } from '../../envs.js';
 
 import HttpError from '../../errors/HttpError.js';
@@ -54,7 +54,7 @@ export default (app) =>
             if (newVersion) {
               try {
                 // eslint-disable-next-line import/no-named-as-default-member
-                await cache.updateVersion(newVersion);
+                await updateVersion(newVersion);
               } catch (error) {
                 throw new HttpError(
                   `Version change: ${error.message}`,
@@ -65,7 +65,7 @@ export default (app) =>
               // Success
               response.status(200).send({
                 statusCode: 200,
-                version: cache.version(),
+                version: version(),
                 message: `Successfully updated Highcharts to version: ${newVersion}.`
               });
             } else {
