+++ lib/server/routes/health.js
@@ -16,7 +16,7 @@ import { readFileSync } from 'fs';
 import { join as pather } from 'path';
 import { log } from '../../logger.js';
 
-import cache from '../../cache.js';
+import { version } from '../../cache.js';
 import { addInterval } from '../../intervals.js';
 import pool from '../../pool.js';
 import { __dirname } from '../../utils.js';
@@ -88,7 +88,7 @@ export default function addHealthRoutes(app) {
           (new Date().getTime() - serverStartTime.getTime()) / 1000 / 60
         ) + ' minutes',
       version: pkgFile.version,
-      highchartsVersion: cache.version(),
+      highchartsVersion: version(),
       averageProcessingTime: stats.spentAverage,
       performedExports: stats.performedExports,
       failedExports: stats.droppedExports,
